Update to v0.1.0-rc2 (#40)
* Handle IOException separately in ApiProcedures

Caught IOException explicitly to provide more granular error handling. This allows to throw the IOException immediately while preserving the behavior for general exceptions. Adjusted throttler release comment to remove redundant text.

* Enable output redirection in ApiProcedures

This commit updates the process start information in ApiProcedures.cs to redirect standard output and error. It also sets UseShellExecute to false and CreateNoWindow to true for better process control and visibility.

* Merge pull request #37

* Add support for bug tracking and enhance profile management
* Add ability to change installation directory dynamically

* Enhance path handling with Path.GetFullPath()

Updated the SystemIOProcedures helper to use Path.GetFullPath for constructing and validating paths. This ensures absolute references, thereby enhancing path accuracy and security checks when manipulating directories and files.

* Add OAuth token authentication method

Implement a new authentication method using OAuth tokens across several classes to enhance security. This method allows users to authenticate using an access token, improving the flexibility and modernizing the authentication process.

* Add ProcessHelper to handle process module monitoring

Introduce ProcessHelper with a StartWatch method to observe and log module changes of a given process. This implementation uses reactive extensions to handle the monitoring and provides module addition and removal logging.

* Prevent overwriting files larger than 100MB.

Added a size check to skip overwriting files that are 100MB or larger in the `SystemIOProcedures` class. This change ensures large files are not unnecessarily replaced, optimizing performance and avoiding potential issues with large file handling.

* Fix potential crash in ProfileFileWatcher process termination

Ensure the process has not exited before attempting to kill it to prevent an InvalidOperationException. Added a try-catch block to safely handle exceptions if the process termination fails.

* Add README for Gml.Client setup and development

Created a comprehensive README guide for the Gml.Client project. It includes step-by-step instructions for cloning, setting up, building, running, and publishing the project, as well as contributing guidelines and additional resources.

---------

+++ src/Gml.Client/Helpers/ProfileFileWatcher.cs
@@ -52,9 +52,16 @@ public class ProfileFileWatcher
 
         FileAdded?.Invoke(this, e.FullPath);
 
-        if (_needKill)
+        if (_needKill && !_process.HasExited)
         {
-            _process.Kill();
+            try
+            {
+                _process.Kill();
+            }
+            catch (InvalidOperationException ex)
+            {
+                // Ignore
+            }
         }
     }
 }
